requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
python-jose
passlib
httpx
redis
langchain
langchain-community

sentence-transformers
fpdf
langdetect
gtts
whisper
torch
fastapi-limiter
